Simplified reify-interval using destructuring.
Decided not to use metadata to pass parameters after all -- this essentially becomes a poor man's environment model, and is inelegant as such. No need to use metadata to indicate the tree is changed: just check for equality to previous tree!

Changes to terminators:
- metadata is now namespace-qualified
- check for terminator? included in terminator-number

Added terminator-like "end-symbols" which replace :# for intervals. Still need to modify current-end so that end-symbols can be used to freeze ends.