Now fully supports end-freezing for arbitrary numbers of strings. This means that once the Ukkonen algorithm is finished, it will be constructing generalized suffix trees, not just regular suffix trees of several serialized strings.
The ends map is now a sorted map, which allows for finding the correct end symbol to use. No new end symbols are generated during the algorithm's run -- it figures out where the ends of the strings are before it runs.

Because ends is a special sorted map, it's necessary to say (into ends {...}) rather than swapping those arguments, to preserve the type.

Printing DOT graphs no longer names them anything. This is unnecessary.

New procedure debug-dot-tree forces the algorithm to run in verbose debug mode even if the debug variable is set to false.